Clean up and fix minnor issues
Test: Existing
